+++ scripts/build-powershell/New-C8yPowershellArguments.ps1
@@ -66,6 +66,7 @@
         "dateto" { "string" }
         "directory" { "string" }
         "file" { "string" }
+        "float" { "float" }
         "fileContents" { "string" }
         "attachment" { "string" }
         "id" { "object" }