+++ tools/PSc8y/Public/Get-BulkOperation.ps1
@@ -1,0 +1,97 @@
+﻿# Code generated from specification version 1.0.0: DO NOT EDIT
+Function Get-BulkOperation {
+<#
+.SYNOPSIS
+Get bulk operation/s
+
+.DESCRIPTION
+Get bulk operation/s
+
+.EXAMPLE
+PS> Get-BulkOperation -Id $BulkOp.id
+
+Get bulk operation by id
+
+
+#>
+    [cmdletbinding(SupportsShouldProcess = $true,
+                   PositionalBinding=$true,
+                   HelpUri='',
+                   ConfirmImpact = 'None')]
+    [Alias()]
+    [OutputType([object])]
+    Param(
+        # Bulk Operation id (required)
+        [Parameter(Mandatory = $true,
+                   ValueFromPipeline=$true,
+                   ValueFromPipelineByPropertyName=$true)]
+        [long]
+        $Id,
+
+        # Show the full (raw) response from Cumulocity including pagination information
+        [Parameter()]
+        [switch]
+        $Raw,
+
+        # Write the response to file
+        [Parameter()]
+        [string]
+        $OutputFile,
+
+        # Ignore any proxy settings when running the cmdlet
+        [Parameter()]
+        [switch]
+        $NoProxy,
+
+        # Specifiy alternative Cumulocity session to use when running the cmdlet
+        [Parameter()]
+        [string]
+        $Session,
+
+        # TimeoutSec timeout in seconds before a request will be aborted
+        [Parameter()]
+        [double]
+        $TimeoutSec
+    )
+
+    Begin {
+        $Parameters = @{}
+        if ($PSBoundParameters.ContainsKey("OutputFile")) {
+            $Parameters["outputFile"] = $OutputFile
+        }
+        if ($PSBoundParameters.ContainsKey("NoProxy")) {
+            $Parameters["noProxy"] = $NoProxy
+        }
+        if ($PSBoundParameters.ContainsKey("Session")) {
+            $Parameters["session"] = $Session
+        }
+        if ($PSBoundParameters.ContainsKey("TimeoutSec")) {
+            $Parameters["timeout"] = $TimeoutSec * 1000
+        }
+
+        if ($env:C8Y_DISABLE_INHERITANCE -ne $true) {
+            # Inherit preference variables
+            Use-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState
+        }
+    }
+
+    Process {
+        foreach ($item in (PSc8y\Expand-Id $Id)) {
+            if ($item) {
+                $Parameters["id"] = if ($item.id) { $item.id } else { $item }
+            }
+
+
+            Invoke-ClientCommand `
+                -Noun "bulkOperations" `
+                -Verb "get" `
+                -Parameters $Parameters `
+                -Type "application/vnd.com.nsn.cumulocity.bulkoperation+json" `
+                -ItemType "" `
+                -ResultProperty "" `
+                -Raw:$Raw
+        }
+    }
+
+    End {}
+}