+++ tools/PSc8y/Public/New-MicroserviceBinary.ps1
@@ -115,21 +115,8 @@ Upload microservice binary
         }
 
         if ($env:C8Y_DISABLE_INHERITANCE -ne $true) {
-            # Inherit preference automatic variables
-            if (!$WhatIfPreference.IsPresent) {
-                $WhatIfPreference = $PSCmdlet.SessionState.PSVariable.get("WhatIfPreference").Value
-            }
-        
-            # Inherit custom parameters
-            $Stack = Get-PSCallStack | Select-Object -Skip 1 -First 1
-            $InheritVariables = @(@{Source="Force"; Target="Force"}, @{Source="WhatIf"; Target="WhatIfPreference"})
-            foreach ($iVariable in $InheritVariables) {
-                if (-Not $PSBoundParameters.ContainsKey($iVariable.Source)) {
-                    if ($null -ne $Stack -and $Stack.InvocationInfo.BoundParameters.ContainsKey($iVariable.Source)) {
-                        Set-Variable -Name $iVariable.Target -Value $Stack.InvocationInfo.BoundParameters[$iVariable.Source] -WhatIf:$false
-                    }
-                }
-            }
+            # Inherit preference variables
+            Use-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState
         }
     }