+++ tools/PSc8y/Public/Remove-BulkOperation.ps1
@@ -1,0 +1,121 @@
+﻿# Code generated from specification version 1.0.0: DO NOT EDIT
+Function Remove-BulkOperation {
+<#
+.SYNOPSIS
+Remove bulk operation/s. Only bulk operations that are in ACTIVE or IN_PROGRESS can be deleted
+
+.DESCRIPTION
+Remove bulk operation/s. Only bulk operations that are in ACTIVE or IN_PROGRESS can be deleted
+
+.EXAMPLE
+PS> Remove-BulkOperation -Id $BulkOp.id
+
+Remove bulk operation by id
+
+
+#>
+    [cmdletbinding(SupportsShouldProcess = $true,
+                   PositionalBinding=$true,
+                   HelpUri='',
+                   ConfirmImpact = 'High')]
+    [Alias()]
+    [OutputType([object])]
+    Param(
+        # Bulk Operation id (required)
+        [Parameter(Mandatory = $true,
+                   ValueFromPipeline=$true,
+                   ValueFromPipelineByPropertyName=$true)]
+        [long]
+        $Id,
+
+        # Cumulocity processing mode
+        [Parameter()]
+        [AllowNull()]
+        [AllowEmptyString()]
+        [ValidateSet("PERSISTENT", "QUIESCENT", "TRANSIENT", "CEP", "")]
+        [string]
+        $ProcessingMode,
+
+        # Show the full (raw) response from Cumulocity including pagination information
+        [Parameter()]
+        [switch]
+        $Raw,
+
+        # Write the response to file
+        [Parameter()]
+        [string]
+        $OutputFile,
+
+        # Ignore any proxy settings when running the cmdlet
+        [Parameter()]
+        [switch]
+        $NoProxy,
+
+        # Specifiy alternative Cumulocity session to use when running the cmdlet
+        [Parameter()]
+        [string]
+        $Session,
+
+        # TimeoutSec timeout in seconds before a request will be aborted
+        [Parameter()]
+        [double]
+        $TimeoutSec,
+
+        # Don't prompt for confirmation
+        [Parameter()]
+        [switch]
+        $Force
+    )
+
+    Begin {
+        $Parameters = @{}
+        if ($PSBoundParameters.ContainsKey("ProcessingMode")) {
+            $Parameters["processingMode"] = $ProcessingMode
+        }
+        if ($PSBoundParameters.ContainsKey("OutputFile")) {
+            $Parameters["outputFile"] = $OutputFile
+        }
+        if ($PSBoundParameters.ContainsKey("NoProxy")) {
+            $Parameters["noProxy"] = $NoProxy
+        }
+        if ($PSBoundParameters.ContainsKey("Session")) {
+            $Parameters["session"] = $Session
+        }
+        if ($PSBoundParameters.ContainsKey("TimeoutSec")) {
+            $Parameters["timeout"] = $TimeoutSec * 1000
+        }
+
+        if ($env:C8Y_DISABLE_INHERITANCE -ne $true) {
+            # Inherit preference variables
+            Use-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState
+        }
+    }
+
+    Process {
+        foreach ($item in (PSc8y\Expand-Id $Id)) {
+            if ($item) {
+                $Parameters["id"] = if ($item.id) { $item.id } else { $item }
+            }
+
+            if (!$Force -and
+                !$WhatIfPreference -and
+                !$PSCmdlet.ShouldProcess(
+                    (PSc8y\Get-C8ySessionProperty -Name "tenant"),
+                    (Format-ConfirmationMessage -Name $PSCmdlet.MyInvocation.InvocationName -InputObject $item)
+                )) {
+                continue
+            }
+
+            Invoke-ClientCommand `
+                -Noun "bulkOperations" `
+                -Verb "delete" `
+                -Parameters $Parameters `
+                -Type "" `
+                -ItemType "" `
+                -ResultProperty "" `
+                -Raw:$Raw
+        }
+    }
+
+    End {}
+}