+++ tools/PSc8y/Public/Update-BulkOperation.ps1
@@ -1,0 +1,153 @@
+﻿# Code generated from specification version 1.0.0: DO NOT EDIT
+Function Update-BulkOperation {
+<#
+.SYNOPSIS
+Update bulk operation
+
+.DESCRIPTION
+Update bulk operation. Making update on a started bulk operation cancels it and creates/schedules a new one.
+
+.EXAMPLE
+PS> Update-BulkOperation -Id $BulkOp.id -CreationRamp 1.5
+
+Update bulk operation wait period between the creation of each operation to 1.5 seconds
+
+
+#>
+    [cmdletbinding(SupportsShouldProcess = $true,
+                   PositionalBinding=$true,
+                   HelpUri='',
+                   ConfirmImpact = 'High')]
+    [Alias()]
+    [OutputType([object])]
+    Param(
+        # Bulk Operation id (required)
+        [Parameter(Mandatory = $true,
+                   ValueFromPipeline=$true,
+                   ValueFromPipelineByPropertyName=$true)]
+        [long]
+        $Id,
+
+        # Delay between every operation creation. (required)
+        [Parameter(Mandatory = $true)]
+        [float]
+        $CreationRampSec,
+
+        # Additional properties to update on the bulk operation.
+        [Parameter()]
+        [object]
+        $Data,
+
+        # Cumulocity processing mode
+        [Parameter()]
+        [AllowNull()]
+        [AllowEmptyString()]
+        [ValidateSet("PERSISTENT", "QUIESCENT", "TRANSIENT", "CEP", "")]
+        [string]
+        $ProcessingMode,
+
+        # Template (jsonnet) file to use to create the request body.
+        [Parameter()]
+        [string]
+        $Template,
+
+        # Variables to be used when evaluating the Template. Accepts a file path, json or json shorthand, i.e. "name=peter"
+        [Parameter()]
+        [string]
+        $TemplateVars,
+
+        # Show the full (raw) response from Cumulocity including pagination information
+        [Parameter()]
+        [switch]
+        $Raw,
+
+        # Write the response to file
+        [Parameter()]
+        [string]
+        $OutputFile,
+
+        # Ignore any proxy settings when running the cmdlet
+        [Parameter()]
+        [switch]
+        $NoProxy,
+
+        # Specifiy alternative Cumulocity session to use when running the cmdlet
+        [Parameter()]
+        [string]
+        $Session,
+
+        # TimeoutSec timeout in seconds before a request will be aborted
+        [Parameter()]
+        [double]
+        $TimeoutSec,
+
+        # Don't prompt for confirmation
+        [Parameter()]
+        [switch]
+        $Force
+    )
+
+    Begin {
+        $Parameters = @{}
+        if ($PSBoundParameters.ContainsKey("CreationRampSec")) {
+            $Parameters["creationRampSec"] = $CreationRampSec
+        }
+        if ($PSBoundParameters.ContainsKey("Data")) {
+            $Parameters["data"] = ConvertTo-JsonArgument $Data
+        }
+        if ($PSBoundParameters.ContainsKey("ProcessingMode")) {
+            $Parameters["processingMode"] = $ProcessingMode
+        }
+        if ($PSBoundParameters.ContainsKey("Template") -and $Template) {
+            $Parameters["template"] = $Template
+        }
+        if ($PSBoundParameters.ContainsKey("TemplateVars") -and $TemplateVars) {
+            $Parameters["templateVars"] = $TemplateVars
+        }
+        if ($PSBoundParameters.ContainsKey("OutputFile")) {
+            $Parameters["outputFile"] = $OutputFile
+        }
+        if ($PSBoundParameters.ContainsKey("NoProxy")) {
+            $Parameters["noProxy"] = $NoProxy
+        }
+        if ($PSBoundParameters.ContainsKey("Session")) {
+            $Parameters["session"] = $Session
+        }
+        if ($PSBoundParameters.ContainsKey("TimeoutSec")) {
+            $Parameters["timeout"] = $TimeoutSec * 1000
+        }
+
+        if ($env:C8Y_DISABLE_INHERITANCE -ne $true) {
+            # Inherit preference variables
+            Use-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState
+        }
+    }
+
+    Process {
+        foreach ($item in (PSc8y\Expand-Id $Id)) {
+            if ($item) {
+                $Parameters["id"] = if ($item.id) { $item.id } else { $item }
+            }
+
+            if (!$Force -and
+                !$WhatIfPreference -and
+                !$PSCmdlet.ShouldProcess(
+                    (PSc8y\Get-C8ySessionProperty -Name "tenant"),
+                    (Format-ConfirmationMessage -Name $PSCmdlet.MyInvocation.InvocationName -InputObject $item)
+                )) {
+                continue
+            }
+
+            Invoke-ClientCommand `
+                -Noun "bulkOperations" `
+                -Verb "update" `
+                -Parameters $Parameters `
+                -Type "application/vnd.com.nsn.cumulocity.bulkoperation+json" `
+                -ItemType "" `
+                -ResultProperty "" `
+                -Raw:$Raw
+        }
+    }
+
+    End {}
+}